Time: 3 ms (39.30%), Space: 6 MB (58.29%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -2,11 +2,12 @@ class Solution {
 public:
     double myPow(double x, int n) {
         double ans = 1.0;
-        long long int nn = n;
+        long long nn = n;
         
         if(nn<0) nn = -1 * nn;
+        
         while(nn){
-            if(nn % 2 == 0){
+            if(nn%2 == 0){
                 x = x * x;
                 nn = nn/2;
             }else{
@@ -14,7 +15,7 @@ public:
                 nn = nn - 1;
             }
         }
-        if(n<0) ans = (double)1.0 / (double)ans;
+        if(n<0) ans = 1/ans;
         return ans;
     }
 };
